14 REST Framework/requirements.txt
django==4.0.2
djangorestframework
